london-bet/resolver/requirements.txt
affine==2.4.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contextily==1.6.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
fiona==1.10.1
fonttools==4.54.1
geographiclib==2.0
geopandas==0.14.4
geopy==2.4.1
h3==3.7.7
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
mercantile==1.2.1
networkx==3.3
numpy==1.26.4
osmnx==1.9.4
packaging==24.1
pandas==2.2.2
pillow==11.0.0
pyarrow==16.1.0
pycparser==2.22
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rasterio==1.4.1
requests==2.32.3
shapely==2.0.6
six==1.16.0
tzdata==2024.1
urllib3==2.2.3
xyzservices==2024.9.0
